ovphysx/tests/python_samples_extra/visual_rerunio_sample/pyproject.toml
[project]
name = "ovphysx-visual-rerunio-sample"
version = "0.0.1"
requires-python = ">=3.10"
dependencies = [
  "ovphysx",
  "numpy",
  "rerun-sdk>=0.29",
]
